add build.py script
At the moment, it generates DSL dictionaries in UTF8 and UTF16 for all data files.
It could be made generate other release outputs too, but I don't need the other
stuff, so it does not.

Meanwhile, the data in `ipa-dict-dsl` repository is 7 years old and it
has a small subset of the dictionaries, so I had to make this.